requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=7.0.0
pypdf>=3.17.1
docarray>=0.39.1
pandas>=2.0.0
pydantic>=2.0.0
pyarrow>=14.0.0
